server/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.10
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==1.4.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
flask-sqlalchemy
pymysql
flask_pymongo
pillow
ultralytics